adds querying for courses in orietations

+++ src/main/java/me/amplitudo/elearning/service/CourseService.java
@@ -1,9 +1,16 @@
 package me.amplitudo.elearning.service;
 
 import me.amplitudo.elearning.domain.Course;
+import me.amplitudo.elearning.domain.User;
+import me.amplitudo.elearning.domain.Year;
 import me.amplitudo.elearning.repository.CourseRepository;
+import me.amplitudo.elearning.repository.UserRepository;
+import me.amplitudo.elearning.repository.YearRepository;
 import me.amplitudo.elearning.service.dto.CourseDTO;
 import me.amplitudo.elearning.service.mapper.CourseMapper;
+import me.amplitudo.elearning.web.rest.errors.BadActionException;
+import me.amplitudo.elearning.web.rest.errors.EntityNotFoundException;
+import me.amplitudo.elearning.web.rest.errors.ExceptionErrors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,6 +19,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Instant;
 import java.util.Optional;
 
 /**
@@ -27,9 +35,17 @@ public class CourseService {
 
     private final CourseMapper courseMapper;
 
-    public CourseService(CourseRepository courseRepository, CourseMapper courseMapper) {
+    private final UserRepository userRepository;
+    private final YearRepository yearRepository;
+
+    public CourseService(CourseRepository courseRepository,
+                         CourseMapper courseMapper,
+                         UserRepository userRepository,
+                         YearRepository yearRepository) {
         this.courseRepository = courseRepository;
         this.courseMapper = courseMapper;
+        this.userRepository = userRepository;
+        this.yearRepository = yearRepository;
     }
 
     /**
@@ -40,8 +56,49 @@ public class CourseService {
      */
     public CourseDTO save(CourseDTO courseDTO) {
         log.debug("Request to save Course : {}", courseDTO);
+
+        Optional<User> professor = Optional.ofNullable(userRepository.findById(courseDTO.getProfessorId())
+            .orElseThrow(() -> new EntityNotFoundException(
+                "Professor with id " + courseDTO.getProfessorId() + " does not exist."
+            )));
+
+        Optional<User> assistant = Optional.ofNullable(userRepository.findById(courseDTO.getAssistantId())
+            .orElseThrow(() -> new EntityNotFoundException(
+                "Assistant with id " + courseDTO.getAssistantId() + " does not exist."
+            )));
+
+        Optional<Year> year = Optional.ofNullable(yearRepository.findById(courseDTO.getYearId())
+            .orElseThrow(() -> new EntityNotFoundException(
+                "Year with id " + courseDTO.getYearId() + " does not exist."
+            )));
+
+        if(courseDTO.getId() != null && !courseRepository.existsById(courseDTO.getId())){
+            throw new EntityNotFoundException(
+                "Course with id " + courseDTO.getId() + " does not exist."
+            );
+        }
+
+        if(
+            courseDTO.getId() == null && courseRepository.countAllByName(courseDTO.getName()) > 0 ||
+            courseDTO.getId() != null && courseRepository.countAllByNameAndIdNot(courseDTO.getName(), courseDTO.getId()) > 0
+        )
+        {
+            throw new BadActionException(
+                ExceptionErrors.COURSE_EXISTS.getErrorCode(),
+                ExceptionErrors.COURSE_EXISTS.getErrorDescription()
+            );
+        }
+
         Course course = courseMapper.toEntity(courseDTO);
+
+        if(course.getId() == null) {
+            course.dateCreated(Instant.now());
+        }else{
+            course.dateUpdated(Instant.now());
+        }
+
         course = courseRepository.save(course);
+
         return courseMapper.toDto(course);
     }
 
@@ -88,6 +145,10 @@ public class CourseService {
      */
     public void delete(Long id) {
         log.debug("Request to delete Course : {}", id);
+        Optional<Course> courseOptional = courseRepository.findById(id);
+        if(!courseOptional.isPresent()){
+            throw new EntityNotFoundException("Course with id " + id + " does not exist.");
+        }
         courseRepository.deleteById(id);
     }
 }
